requirements.txt
python-dotenv>=1.0.1
numpy>=1.26.4
scikit-learn>=1.5.1
datasets>=2.20.0
tqdm>=4.66.4
streamlit>=1.44.0
langchain>=0.3.23
langchain-community>=0.3.21
langchain-ollama>=0.3.2
chromadb>=0.5.5
